+++ docs/plans/5.5-6.0-resource-translations-api.md
@@ -209,7 +209,7 @@ The **read** side stays as a convenience: the admin `ProductSerializer` exposes 
 module Spree::Api::V3::Admin
   class ProductSerializer < V3::ProductSerializer
     attribute :translations, if: proc { expand?('translations') } do |product|
-      Spree::Translations::Matrix.for(product) # { "de" => { "name" => …, "translated_field_count" => 5 }, … }
+      Spree::Translations.matrix_for(product) # { "de" => { "name" => …, "translated_field_count" => 5 }, … }
     end
   end
 end
