+++ spree/core/lib/spree/core.rb
@@ -494,6 +494,7 @@ require 'spree/core/engine'
 require 'spree/i18n'
 require 'spree/localized_names'
 require 'spree/localized_number'
+require 'spree/translations'
 require 'spree/money'
 require 'spree/permitted_attributes'
 require 'spree/service_module'
